Added numberOfEnemiesToReleaseAtATime and MaxNumberofEnemiesOnScreenAtATime to level settings.
Created a stats manager to display player health, score and various other stats.

Tracking stats of the number of enemies that you have encountered and the number of enemies that have been killed as well as your kill percentage. At the moment the I am displaying all this information on the screen through the stats manager.